AgenticEnv/requirements.txt
# ========================================
# Core Framework Dependencies
# ========================================
langgraph>=0.2.0
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0

# ========================================
# LLM Provider
# ========================================
groq>=0.11.0
langchain-groq>=0.2.0

# ========================================
# Database & Storage
# ========================================
# SQLite comes with Python, but these help with ORM-like features
sqlalchemy>=2.0.0

# ========================================
# Configuration Management
# ========================================
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# ========================================
# Data Processing
# ========================================
pandas>=2.0.0
numpy>=1.24.0

# ========================================
# JSON & Schema Validation
# ========================================
jsonschema>=4.20.0

# ========================================
# Date & Time Utilities
# ========================================
python-dateutil>=2.8.0

# ========================================
# HTTP Requests (for future API integrations)
# ========================================
requests>=2.31.0
httpx>=0.27.0

# ========================================
# Async Support
# ========================================
aiohttp>=3.9.0
asyncio>=3.4.3

# ========================================
# Logging & Monitoring
# ========================================
colorlog>=6.8.0
rich>=13.7.0

# ========================================
# Text Processing
# ========================================
tiktoken>=0.5.0  # Token counting for OpenAI-compatible models

# ========================================
# Testing (Optional - for future development)
# ========================================
pytest>=8.0.0
pytest-asyncio>=0.23.0

# ========================================
# Type Checking (Optional - for development)
# ========================================
mypy>=1.8.0
types-requests>=2.31.0

# ========================================
# Code Quality (Optional - for development)
# ========================================
black>=24.0.0
flake8>=7.0.0
isort>=5.13.0
